Added documentation
darcs-hash:7a8e06aebfa3c99ac72d1ff0c3d67294febbc562

+++ cbits/win32.c
@@ -96,12 +96,12 @@ void *system_io_mmap_mmap(void *handle, int mode, long long offset, int size)
     mapping = CreateFileMapping(handle, NULL, flProtect, (DWORD) ((offset + size)>>32), (DWORD)(offset + size), NULL);
     if( !mapping ) {
       DWORD dw = GetLastError();
-      fprintf(stderr,"CreateFileMapping %d\n",dw);
+      fprintf(stderr,"CreateFileMapping %d\n",(int)dw);
     }
     ptr = MapViewOfFile(mapping,dwDesiredAccess, (DWORD)(offset>>32), (DWORD)(offset), size );
     if( !ptr ) {
       DWORD dw = GetLastError();
-      fprintf(stderr,"MapViewOfFile %d\n",dw);
+      fprintf(stderr,"MapViewOfFile %d\n",(int)dw);
     }
     CloseHandle(mapping);
     return ptr;
@@ -128,3 +128,5 @@ int system_io_mmap_granularity()
     GetSystemInfo(&sysinfo);
     return sysinfo.dwAllocationGranularity;
 }
+
+
